simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
]
